Fix CMake unable to find `SetSystemProcessor` module
The "Modules" directory is included like the root CMakeLists.txt
following:

list (APPEND CMAKE_MODULE_PATH ${CMAKE_SOURCE_DIR}/cmake/Modules)

Which confuses CMake when building soxr as an external
project/sub-project.

CMAKE_SOURCE_DIR points to the root directory of the top-level project,
which is not the same as soxrs root directory when the library is used
as an external project.

CMAKE_CURRENT_SOURCE_DIR points to the source directory (that contains
the CMakeLists.txt file) currently being processed, which points to
soxrs root directory, as that's the intended behavior in this case.

This problem becomes apparent when using soxr as an external project,
either using CMake's ExternalProject module, or simply add_subdirectory.
Because in that case ${CMAKE_SOURCE_DIR} is set to the outer project's
root directory, not soxrs, as a result, the path added to
${CMAKE_MODULE_PATH} is invalid/non-existent, making it impossible to
build the library without manually fixing the CMake script first.

This commit fixes the issue by including the modules directory
relative to CMAKE_CURRENT_SOURCE_DIR instead of CMAKE_SOURCE_DIR